refactor: move file and link in their own files
cc #7
